Initial work towards implementing #2086
At this point, goto based flow control is complete.
So is variable interpretation.
However, expression interpretation is still in an immature state, and
while it is completely functional, it may not be desirable because it
does not follow order of operations and computes from right to left.
Lastly, the methods set, round, floor, ceil, inc, mul, div, mod have
been added for simple operations.
In order to support this Methods have been modified to have access to
variables.
This will break existing addons that add custom methods, but there is
not much I can do about that and still be able to have methods that can
mutate variables.

Lastly, to prevent infinite loops from happening unless the player
intends for an infinite loop, it keeps track of how many time in a row
the same label has been visited, and after 100 visits, (will be
configurable), it kills the program.

Finally, to prevent a program from running forever, if a program has
been running for more than 1000 ms, (will be configurable), it will be
killed.

The only reason these last 2 pieces are important is because forge
essentials performs scripting on the main thread, making background
scripts impossible.