Should Finish #2086
Finished Expression parsing to add functions and order of operations.
Currently nested functions is not possible though you can have
expressions in functions.
Deprecated '@Random' in favor of the new randi and randf functions.
Reverted Script methods to previous state
Improved parsing of variables, arguments, and expressions to be more
efficient.
Added support for expressions in if conditions and in set method.
added loadp, loadg, storep, storeg
(storep and storeg) was added to prevent confusion over where to fetch
variables and to ensure that all variables pointed to fe.vars.<name>
instead of some arbitrary permission that could be overwritten by
accident.